pyproject.toml
[tool.poetry]
name = "idao-21-baseline"
version = "0.1.0"
description = ""
authors = ["Abdalaziz Rashid <abdalaziz.rashid@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
pytorch-lightning = "^1.2.3"
torchvision = "^0.9.0"
matplotlib = "^3.3.4"
scikit-plot = "^0.3.7"
more-itertools = "^8.7.0"
pandas = "^1.2.3"
[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
